adjusted the app to use custom hooks

+++ src/App.js
@@ -1,4 +1,6 @@
 import { useState, useEffect } from 'react';
+import useWindowSize from './hooks/useWindowSize';
+import useAxiosFetch from './hooks/useAxiosFetch';
 import { Routes, Route, useNavigate } from 'react-router-dom';
 import { format } from 'date-fns';
 import api from './api/posts';
@@ -11,6 +13,8 @@ import Missing from './components/Missing';
 import EditPost from './components/EditPost';
 
 const App = () => {
+  const API_URL = 'http://localhost:3500/posts';
+
   const [posts, setPosts] = useState([]);
   const [search, setSearch] = useState('');
   const [searchResults, setSearchResults] = useState([]);
@@ -21,24 +25,13 @@ const App = () => {
 
   const navigate = useNavigate();
 
-  useEffect(() => {
-    const fetchPosts = async () => {
-      try {
-        const response = await api.get('/posts');
-        setPosts(response.data);
-      } catch (err) {
-        if (err.response) {
-          console.log(err.response.data);
-          console.log(err.response.status);
-          console.log(err.response.headers);
-        } else {
-          console.log(`Error: ${err.message}`);
-        }
-      }
-    };
+  const { width } = useWindowSize();
 
-    fetchPosts();
-  }, []);
+  const { data, fetchError, isLoading } = useAxiosFetch(API_URL);
+
+  useEffect(() => {
+    setPosts(data);
+  }, [data]);
 
   useEffect(() => {
     const filteredResults = posts.filter(
@@ -94,13 +87,19 @@ const App = () => {
   };
 
   const handleDelete = async (id) => {
-    try {
-      await api.delete(`/posts/${id}`);
-      const postsList = posts.filter((post) => post.id !== id);
-      setPosts(postsList);
-      navigate('/');
-    } catch (err) {
-      console.log(`Error: ${err.message}`);
+    const confirmed = window.confirm(
+      'Are you sure you want to delete this post?'
+    );
+
+    if (confirmed) {
+      try {
+        await api.delete(`/posts/${id}`);
+        const postsList = posts.filter((post) => post.id !== id);
+        setPosts(postsList);
+        navigate('/');
+      } catch (err) {
+        console.log(`Error: ${err.message}`);
+      }
     }
   };
 
@@ -108,9 +107,18 @@ const App = () => {
     <Routes>
       <Route
         path="/"
-        element={<Layout search={search} setSearch={setSearch} />}
+        element={<Layout search={search} setSearch={setSearch} width={width} />}
       >
-        <Route index element={<Home posts={searchResults} />} />
+        <Route
+          index
+          element={
+            <Home
+              posts={searchResults}
+              fetchError={fetchError}
+              isLoading={isLoading}
+            />
+          }
+        />
         <Route path="post">
           <Route
             index
